refactor move to debug file
